+++ packages/expo-modules-autolinking/src/commands/verifyCommand.ts
@@ -1,16 +1,58 @@
 import chalk from 'chalk';
+import commander from 'commander';
 import fs from 'fs';
 import path from 'path';
 
 import {
-  BaseDependencyResolution,
-  DependencyResolution,
+  AutolinkingCommonArguments,
+  createAutolinkingOptionsLoader,
+  registerAutolinkingArguments,
+} from './autolinkingOptions';
+import {
+  type BaseDependencyResolution,
+  type DependencyResolution,
+  type ResolutionResult,
   DependencyResolutionSource,
-  ResolutionResult,
+  makeCachedDependenciesLinker,
+  mergeResolutionResults,
+  scanDependencyResolutionsForPlatform,
 } from '../dependencies';
 
+interface VerifyArguments extends AutolinkingCommonArguments {
+  verbose?: boolean | null;
+  json?: boolean | null;
+}
+
+export function verifyCommand(cli: commander.CommanderStatic) {
+  return registerAutolinkingArguments(cli.command('verify'))
+    .option('-v, --verbose', 'Output all results instead of just warnings.', () => true, false)
+    .option('-j, --json', 'Output results in the plain JSON format.', () => true, false)
+    .option(
+      '-p, --platform [platform]',
+      'The platform to validate native modules for. Available options: "android", "ios", "both"',
+      'both'
+    )
+    .action(async (commandArguments: VerifyArguments) => {
+      const platforms =
+        commandArguments.platform === 'both' ? ['android', 'ios'] : [commandArguments.platform!];
+      const autolinkingOptionsLoader = createAutolinkingOptionsLoader(commandArguments);
+      const appRoot = await autolinkingOptionsLoader.getAppRoot();
+      const linker = makeCachedDependenciesLinker({ projectRoot: appRoot });
+      const results = mergeResolutionResults(
+        await Promise.all(
+          platforms.map((platform) => scanDependencyResolutionsForPlatform(linker, platform))
+        )
+      );
+      await verifySearchResults(results, {
+        appRoot,
+        verbose: !!commandArguments.verbose,
+        json: !!commandArguments.json,
+      });
+    });
+}
+
 interface VerifyOptions {
-  projectRoot: string;
+  appRoot: string;
   verbose?: boolean;
   json?: boolean;
 }
@@ -29,7 +71,7 @@ export async function verifySearchResults(
   results: ResolutionResult,
   options: VerifyOptions
 ): Promise<void> {
-  const { projectRoot } = options;
+  const { appRoot } = options;
 
   async function getHumanReadableDependency(dependency: BaseDependencyResolution): Promise<string> {
     let version = dependency.version || null;
@@ -47,7 +89,7 @@ export async function verifySearchResults(
         version = null;
       }
     }
-    const relative = path.relative(projectRoot, dependency.originPath);
+    const relative = path.relative(appRoot, dependency.originPath);
     return version
       ? `${dependency.name}@${version} (at: ${relative})`
       : `${dependency.name} at: ${relative}`;